pyproject.toml
[project]
name = "py-image-to-ascii-art"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=12.1.0",
]
